+++ src/app/sessions/[id]/_components/SessionDetail/SessionDetailErrorFallback.tsx
